move UBO_BINDINGS from projection to to shared uniform_buffer file
